requirements.txt
requests
pyautogui
python-telegram-bot==13.7
psutil==5.8.0
pynput==1.7.6
opencv-python==4.5.5.64
sounddevice==0.4.5
scipy==1.8.0
pywin32==303; sys_platform == 'win32'
psutil==5.9.5
pywin32==306
